Moved ResponsiveJText
